requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1
attrs==21.4.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorhash==1.2.1
configparser==5.3.0
cryptography==38.0.1
cssselect2==0.5.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Limiter==2.0.4
flask-marshmallow==0.14.0
Flask-MonitoringDashboard==3.1.1
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
fonttools==4.31.2
google-api-core==2.4.0
google-api-python-client==2.58.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-core==2.2.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
imgkit==1.2.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
limits==2.2.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysql-connector-python==8.0.27
numpy==1.23.4
oauthlib==3.2.0
omise==0.11.0
Pillow==9.0.1
protobuf==3.19.3
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydyf==0.1.2
PyMySQL==1.0.2
pyparsing==3.0.9
pyphen==0.12.0
pyrsistent==0.18.0
python-dotenv==0.19.2
python-http-client==3.3.7
pytz==2021.3
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.9.3
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.29
starkbank-ecdsa==2.2.0
tinycss2==1.1.1
tzlocal==2.0.0
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.2
wkhtmltopdf==0.2
zopfli==0.2.1
geographiclib==2.0
geopy==2.3.0
pyodbc==4.0.32
redis==4.5.5
stripe
openai
python-dateutil==2.8.2
